0.6

+++ add_domain.sh
@@ -1,20 +1,18 @@
 #!/bin/bash
 
 # Copyright 2017-2018 Tim Scharner (https://timscha.io)
-# Version 0.5.0
+# Version 0.6.0
 
-# "Domain: $USER_MAINDOMAIN"
-# "Subdomain $USER_SUBDOMAIN"
-# "Absolute path maindomain: $HOST_MAINDOMAIN_ROOT_LOCATION"
-# "Absolute path subdomain: $HOST_SUBDOMAIN_ROOT_LOCATION"
-# "Absolute path htdocs maindomain: $HOST_MAINDOMAIN_HTTPD_LOCATION"
-# "Absolute path htdocs subdomain: $HOST_SUBDOMAIN_HTTPD_LOCATION"
-# "MySQL username: $HOST_DB_USER"
-# "MySQL password: $HOST_DB_PASS"
-# "MySQL database: $HOST_DB_DATABASE"
-# "Location owner: $HOST_LOCATION_USER"
-# "Domain with hyphen: $USER_DOMAIN_HYPHEN"
-# "Subdomain with hyphen: $USER_SUBDOMAIN_HYPHEN"
+servicesCheck(){
+ps cax | grep $1 > /dev/null
+if [ $? -eq 0 ]; then
+  return 1
+else
+  return 0
+fi
+}
+
+source includes/install_web.sh
 
 if [ -e /etc/redhat-release ]; then
      DISTRO="centos"
@@ -22,9 +20,20 @@ elif [ -e /etc/debian_version ]; then
      DISTRO="debian"
 fi
 
+if [ -e /etc/apache2/apache2.conf ]; then
+     WEBSRV="apache"
+elif [ -e /etc/nginx/nginx.conf ]; then
+     WEBSRV="nginx"
+fi
+
 create_skeleton_dirs() {
 	useradd $HOST_LOCATION_USER -d /var/www/$USER_MAINDOMAIN
+
+  if [ $WEBSRV = "nginx" ]; then
 	usermod -aG $HOST_LOCATION_USER nginx
+  elif [ $WEBSRV = "apache" ]; then
+	usermod -aG $HOST_LOCATION_USER www-data
+  fi
 
   if [ ! -d /var/www/$USER_MAINDOMAIN ]; then
 	   mkdir -p /var/www/$USER_MAINDOMAIN/htdocs
@@ -33,11 +42,11 @@ create_skeleton_dirs() {
   fi
 
   if [ $USER_DOMAIN_TYP = "1" ]; then
-	if [ ! -d /var/www/$USER_MAINDOMAIN/$USER_SUBDOMAIN ]; then
-    mkdir -p /var/www/$USER_MAINDOMAIN/$USER_SUBDOMAIN/htdocs
-		mkdir /var/www/$USER_MAINDOMAIN/$USER_SUBDOMAIN/logs
-		mkdir /var/www/$USER_MAINDOMAIN/$USER_SUBDOMAIN/tmp
-	fi
+  	if [ ! -d /var/www/$USER_MAINDOMAIN/$USER_SUBDOMAIN ]; then
+      mkdir -p /var/www/$USER_MAINDOMAIN/$USER_SUBDOMAIN/htdocs
+  		mkdir /var/www/$USER_MAINDOMAIN/$USER_SUBDOMAIN/logs
+  		mkdir /var/www/$USER_MAINDOMAIN/$USER_SUBDOMAIN/tmp
+  	fi
   fi
 
   chown -R $HOST_LOCATION_USER: /var/www/$USER_MAINDOMAIN
@@ -167,10 +176,73 @@ configure_fpm_pool(){
         sed -i 's|'HOST_DOMAIN_FULL'|'$HOST_SUBDOMAIN_ROOT_LOCATION'|g' /etc/opt/remi/php70/php-fpm.d/$USER_SUBDOMAIN.conf
 		    sed -i s/PHP-SOCKET/php70-fpm-$USER_SUBDOMAIN_HYPHEN/g /etc/opt/remi/php70/php-fpm.d/$USER_MAINDOMAIN.conf
       fi
-      systemctl reload php70-php-fpm
+      systemctl -q reload php70-php-fpm
     fi
   fi
 	return 0
+}
+
+configure_apache_vhost() {
+  if [ $USER_DOMAIN_TYP = "0" ]; then
+    if [ $USER_PHP_VERSION = "php72" ]; then
+      NGXSOCKET="/var/run/php72-fpm-$USER_DOMAIN_HYPHEN.sock"
+    fi
+    if [ $USER_PHP_VERSION = "php71" ]; then
+      NGXSOCKET="/var/run/php71-fpm-$USER_DOMAIN_HYPHEN.sock"
+    fi
+    if [ $USER_PHP_VERSION = "php70" ]; then
+      NGXSOCKET="/var/run/php70-fpm-$USER_DOMAIN_HYPHEN.sock"
+    fi
+    if [ $USER_CMS_CHOICE = "none" ]; then
+      cp templates/apache_default.template /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+    fi
+    if [ $USER_CMS_CHOICE = "nextcloud" ]; then
+      cp templates/apache_nextcloud.template /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+    fi
+    if [ $USER_CMS_CHOICE = "wordpress" ]; then
+      cp templates/apache_wordpress.template /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+    fi
+
+    sed -i s/DOMAIN_HYPHEN/$USER_DOMAIN_HYPHEN/g /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+    sed -i 's|'NGXSOCKET'|'$NGXSOCKET'|g' /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+    sed -i s/DOMAIN_FULLNAME/$USER_MAINDOMAIN/g /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+    sed -i s/SSL_DOMAINNAME_FULLNAME/$USER_MAINDOMAIN/g /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+    sed -i 's|'DOMAIN_HTTPD_LOCATION'|'$HOST_MAINDOMAIN_HTTPD_LOCATION'|g' /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+    sed -i 's|'HOST_ROOT_LOCATION'|'$HOST_MAINDOMAIN_ROOT_LOCATION'|g' /etc/apache2/sites-available/$USER_MAINDOMAIN.conf
+
+    a2ensite -q $USER_MAINDOMAIN.conf
+  fi
+
+  if [ $USER_DOMAIN_TYP = "1" ]; then
+    if [ $USER_PHP_VERSION = "php72" ]; then
+      NGXSOCKET="/var/run/php72-fpm-$USER_SUBDOMAIN_HYPHEN.sock;"
+    fi
+    if [ $USER_PHP_VERSION = "php71" ]; then
+      NGXSOCKET="/var/run/php71-fpm-$USER_SUBDOMAIN_HYPHEN.sock;"
+    fi
+    if [ $USER_PHP_VERSION = "php70" ]; then
+      NGXSOCKET="/var/run/php70-fpm-$USER_SUBDOMAIN_HYPHEN.sock;"
+    fi
+    if [ $USER_CMS_CHOICE = "none" ]; then
+      cp templates/apache_default.template /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+    fi
+    if [ $USER_CMS_CHOICE = "nextcloud" ]; then
+      cp templates/apache_nextcloud.template /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+    fi
+    if [ $USER_CMS_CHOICE = "wordpress" ]; then
+      cp templates/apache_default.template /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+    fi
+
+    sed -i s/DOMAIN_HYPHEN/$USER_DOMAIN_HYPHEN/g /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+    sed -i 's|'NGXSOCKET'|'$NGXSOCKET'|g' /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+    sed -i s/DOMAIN_FULLNAME/$USER_MAINDOMAIN/g /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+    sed -i s/SSL_DOMAINNAME_FULLNAME/$USER_MAINDOMAIN/g /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+    sed -i 's|'DOMAIN_HTTPD_LOCATION'|'$HOST_MAINDOMAIN_HTTPD_LOCATION'|g' /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+    sed -i 's|'HOST_ROOT_LOCATION'|'$HOST_MAINDOMAIN_ROOT_LOCATION'|g' /etc/apache2/sites-available/$USER_SUBDOMAIN.conf
+
+    a2ensite -q $USER_SUBDOMAIN.conf
+  fi
+  return 0
 }
 
 configure_nginx_vhost(){
@@ -184,7 +256,15 @@ configure_nginx_vhost(){
     if [ $USER_PHP_VERSION = "php70" ]; then
       NGXSOCKET="/var/run/php70-fpm-$USER_DOMAIN_HYPHEN.sock;"
     fi
-    cp templates/nginx_default.template /etc/nginx/conf.d/$USER_MAINDOMAIN.conf
+    if [ $USER_CMS_CHOICE = "none" ]; then
+      cp templates/nginx_default.template /etc/nginx/conf.d/$USER_MAINDOMAIN.conf
+    fi
+    if [ $USER_CMS_CHOICE = "nextcloud" ]; then
+      cp templates/nginx_nextcloud.template /etc/nginx/conf.d/$USER_MAINDOMAIN.conf
+    fi
+    if [ $USER_CMS_CHOICE = "wordpress" ]; then
+      cp templates/nginx_wordpress.template /etc/nginx/conf.d/$USER_MAINDOMAIN.conf
+    fi
     sed -i s/DOMAIN_HYPHEN/$USER_DOMAIN_HYPHEN/g /etc/nginx/conf.d/$USER_MAINDOMAIN.conf
     sed -i 's|'NGXSOCKET'|'$NGXSOCKET'|g' /etc/nginx/conf.d/$USER_MAINDOMAIN.conf
     sed -i s/DOMAIN_FULLNAME/$USER_MAINDOMAIN/g /etc/nginx/conf.d/$USER_MAINDOMAIN.conf
@@ -194,15 +274,23 @@ configure_nginx_vhost(){
   fi
   if [ $USER_DOMAIN_TYP = "1" ]; then
     if [ $USER_PHP_VERSION = "php72" ]; then
-      NGXSOCKET="/var/run/php72-fpm-$USER_SUBDOMAIN.sock;"
+      NGXSOCKET="/var/run/php72-fpm-$USER_SUBDOMAIN_HYPHEN.sock;"
     fi
     if [ $USER_PHP_VERSION = "php71" ]; then
-      NGXSOCKET="/var/run/php71-fpm-$USER_SUBDOMAIN.sock;"
+      NGXSOCKET="/var/run/php71-fpm-$USER_SUBDOMAIN_HYPHEN.sock;"
     fi
     if [ $USER_PHP_VERSION = "php70" ]; then
-      NGXSOCKET="/var/run/php70-fpm-$USER_SUBDOMAIN.sock;"
+      NGXSOCKET="/var/run/php70-fpm-$USER_SUBDOMAIN_HYPHEN.sock;"
     fi
-    cp templates/nginx_default.template /etc/nginx/conf.d/$USER_SUBDOMAIN.conf
+    if [ $USER_CMS_CHOICE = "none" ]; then
+      cp templates/nginx_default.template /etc/nginx/conf.d/$USER_SUBDOMAIN.conf
+    fi
+    if [ $USER_CMS_CHOICE = "nextcloud" ]; then
+      cp templates/nginx_nextcloud.template /etc/nginx/conf.d/$USER_SUBDOMAIN.conf
+    fi
+    if [ $USER_CMS_CHOICE = "wordpress" ]; then
+      cp templates/nginx_wordpress.template /etc/nginx/conf.d/$USER_SUBDOMAIN.conf
+    fi
     sed -i s/DOMAIN_HYPHEN/$USER_SUBDOMAIN_HYPHEN/g /etc/nginx/conf.d/$USER_SUBDOMAIN.conf
     sed -i 's|'NGXSOCKET'|'$NGXSOCKET'|g' /etc/nginx/conf.d/$USER_SUBDOMAIN.conf
     sed -i s/DOMAIN_FULLNAME/$USER_SUBDOMAIN/g /etc/nginx/conf.d/$USER_SUBDOMAIN.conf
@@ -213,8 +301,7 @@ configure_nginx_vhost(){
 	return 0
 }
 
-configure_logrorate_nginx() {
-  # Logrotation for nginx
+configure_logrotate() {
   if [ $USER_DOMAIN_TYP = "0" ]; then
     cat >> /etc/logrotate.d/nginx <<EOL
     /var/log/www/$USER_MAINDOMAIN/logs/*.log {
@@ -252,14 +339,26 @@ EOL
   return 0
 }
 
-configure_letsencrypt_nginx() {
-  if [ $USER_DOMAIN_TYP = "0" ]; then
-    certbot certonly --standalone --agree-tos --email hostmaster@$USER_MAINDOMAIN --pre-hook "systemctl stop nginx" --post-hook "systemctl start nginx" --rsa-key-size 4096 -d $USER_MAINDOMAIN -d www.$USER_MAINDOMAIN
-    return 0
+configure_letsencrypt() {
+  if [ $WEBSRV = "nginx" ]; then
+    if [ $USER_DOMAIN_TYP = "0" ]; then
+      certbot certonly --standalone --agree-tos --email hostmaster@$USER_MAINDOMAIN --pre-hook "systemctl stop nginx" --post-hook "systemctl start nginx" --rsa-key-size 4096 -d $USER_MAINDOMAIN -d www.$USER_MAINDOMAIN
+      return 0
+    fi
+    if [ $USER_DOMAIN_TYP = "1" ]; then
+      certbot certonly --standalone --agree-tos --email hostmaster@$USER_MAINDOMAIN --pre-hook "systemctl stop nginx" --post-hook "systemctl start nginx" --rsa-key-size 4096 -d $USER_MAINDOMAIN -d www.$USER_MAINDOMAIN -d $USER_SUBDOMAIN
+      return 0
+    fi
   fi
-  if [ $USER_DOMAIN_TYP = "1" ]; then
-    certbot certonly --standalone --agree-tos --email hostmaster@$USER_MAINDOMAIN --pre-hook "systemctl stop nginx" --post-hook "systemctl start nginx" --rsa-key-size 4096 -d $USER_MAINDOMAIN -d www.$USER_MAINDOMAIN -d $USER_SUBDOMAIN
-    return 0
+  if [ $WEBSRV = "apache" ]; then
+    if [ $USER_DOMAIN_TYP = "0" ]; then
+      certbot certonly --standalone --agree-tos --email hostmaster@$USER_MAINDOMAIN --pre-hook "systemctl stop apache2" --post-hook "systemctl start apache2" --rsa-key-size 4096 -d $USER_MAINDOMAIN -d www.$USER_MAINDOMAIN
+      return 0
+    fi
+    if [ $USER_DOMAIN_TYP = "1" ]; then
+      certbot certonly --standalone --agree-tos --email hostmaster@$USER_MAINDOMAIN --pre-hook "systemctl stop apache2" --post-hook "systemctl start apache2" --rsa-key-size 4096 -d $USER_MAINDOMAIN -d www.$USER_MAINDOMAIN -d $USER_SUBDOMAIN
+      return 0
+    fi
   fi
 }
 
@@ -288,14 +387,6 @@ EOSQL
 
 echo <<EOF "
 $(basename $0) will attempt to add a vhost to your system now.
-
--------------------------------------------------------------------------------
-For Let's Encrypt, please read the Terms of Service at
-https://letsencrypt.org/documents/LE-SA-v1.2-November-15-2017.pdf.
-With using this script you agree in order to register with the ACME server at
-https://acme-v01.api.letsencrypt.org/directory
--------------------------------------------------------------------------------
-
 This script is provided as it is, no warraties implied. (Ctrl-c to abort)
 "
 EOF
@@ -352,17 +443,42 @@ then
     esac
   done
 
-  read -p "Do you want to add a database? [y/n]" -n 1 -r
-  echo
-  if [[ $REPLY =~ ^[Yy]$ ]]
-  then
-    printf "Please provide your MySQL-Root-Password when asked."
+  PS3='Select an Installer for your domain: '
+  options=("Nextcloud" "Wordpress" "None")
+  select opt in "${options[@]}"
+  do
+    case $opt in
+      "Nextcloud")
+        echo "Nextcloud selected"
+        USER_CMS_CHOICE=nextcloud
+        break
+        ;;
+      "Wordpress")
+        echo "Wordpress selected"
+        USER_CMS_CHOICE=wordpress
+        break
+        ;;
+      "None")
+        USER_CMS_CHOICE=none
+        break
+        ;;
+      *) echo invalid option;;
+    esac
+  done
+
+  if [ $USER_CMS_CHOICE = "nextcloud" ] || [ $USER_CMS_CHOICE = "wordpress" ]; then
     USER_DB_SITE=1
   else
-    USER_DB_SITE=0
+    read -p "Do you want to add a database? [y/n] " -n 1 -r
+    echo
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+      printf "\nPlease provide your MySQL-Root-Password when asked.\n"
+      USER_DB_SITE=1
+    else
+      USER_DB_SITE=0
+    fi
   fi
-
-  # Do we got everything we need?
 
   HOST_LOCATION_USER=(${USER_MAINDOMAIN//./ })
 
@@ -387,11 +503,21 @@ then
 
   create_skeleton_dirs
   configure_fpm_pool
-  configure_nginx_vhost
-  configure_logrorate_nginx
-  configure_letsencrypt_nginx
+  if [ $WEBSRV = "nginx" ]; then
+    configure_nginx_vhost
+  elif [ $WEBSRV = "apache" ]; then
+    configure_apache_vhost
+  fi
+  configure_logrotate
+  configure_letsencrypt
   if [ $USER_DB_SITE = "1" ]; then
   configure_database
+  fi
+
+  if [ $USER_CMS_CHOICE = "nextcloud" ]; then
+    configure_nextcloud
+  elif [ $USER_CMS_CHOICE = "wordpress" ]; then
+    configure_wordpress
   fi
 
   [ $? -ne "0" ] && exit 1
@@ -402,12 +528,21 @@ echo <<EOF "
 #
 # Your domain is ready to use:
 #
-# Domain: $USER_MAINDOMAIN
-# Absolute path: $HOST_MAINDOMAIN_ROOT_LOCATION
-# MySQL username: $HOST_DB_USER
-# MySQL password: $HOST_DB_PASS
-# MyMySQL database: $HOST_DB_DATABASE
-# Location owner: $HOST_LOCATION_USER
+"
+EOF
+echo "Domain: $USER_MAINDOMAIN"
+if [ $USER_DOMAIN_TYP = "1" ]; then
+  echo "Subdomain: $USER_SUBDOMAIN"
+fi
+echo "Absolute path domain: $HOST_MAINDOMAIN_ROOT_LOCATION"
+if [ $USER_DOMAIN_TYP = "1" ]; then
+  echo "Absolute path subdomain: $HOST_SUBDOMAIN_ROOT_LOCATION"
+fi
+echo "MySQL username: $HOST_DB_USER"
+echo "MySQL password: $HOST_DB_PASS"
+echo "MyMySQL database: $HOST_DB_DATABASE"
+echo "Location owner: $HOST_LOCATION_USER"
+echo <<EOF "
 #
 #
 #################################################################
